\name{rvkstat-package}
\alias{rvkstat-package}
\alias{rvkstat}
\docType{package}
\title{
\packageTitle{rvkstat}
}
\description{
\packageDescription{rvkstat}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rvkstat}
\packageIndices{rvkstat}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{rvkstat}

Maintainer: \packageMaintainer{rvkstat}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
