% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vkGetDbUniversities.R
\name{vkGetDbUniversities}
\alias{vkGetDbUniversities}
\title{Returns a List of Universities}
\usage{
vkGetDbUniversities(
  q = NULL,
  country_id = NULL,
  city_id = NULL,
  username = getOption("rvkstat.username"),
  api_version = getOption("rvkstat.api_version"),
  token_path = vkTokenPath(),
  access_token = getOption("rvkstat.access_token")
)
}
\arguments{
\item{q}{Search query string.}

\item{country_id}{Country ID, \code{\link{vkGetDbCountries}}}

\item{city_id}{City ID, \code{\link{vkGetDbCities}}}

\item{username}{Your vkontakte login.}

\item{api_version}{Vkontakte API version.}

\item{token_path}{Path to dir with credentials}

\item{access_token}{API access tokens obtained using vkAuth or vkGetToken functions}
}
\value{
tibble with Universities list
}
\description{
Returns a List of Universities
}
