\name{vkLoadToken}
\alias{vkLoadToken}
\title{Load vkontakte developer token}
\description{vkLoadToken load your developer token, which you need for work with vk.com api }
\usage{
vkLoadToken()
}
\value{R object with vk.com token}
\author{Alexey Seleznev}
\examples{
##---- Get token from work directory
myToken <- vkLoadToken()
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
