\name{vkGetAds}
\alias{vkGetAds}
\title{Получение списка объявлений из Вконтакте}
\description{Функция vkGetAds даёт возможность получить список рекламных объявлений из рекламного кабинета Вконтакте}
\usage{
vkGetAds(account_id = NULL, client_id = NULL, include_deleted = TRUE, campaign_ids = "null", ad_ids = "null", access_token = NULL))
}
\arguments{
  \item{account_id}{Идентификатор рекламного кабинета, список всех доступных рекламных кабинетов можно получить с помщью функции vkgetAccounts.}
  \item{client_id}{Идентификатор клиента, у которого запрашиваются рекламные объявления. Доступно и обязательно для рекламных агентств.}
  \item{include_deleted}{Логическое TRUE или FALSE, флаг, задающий необходимость вывода архивных объявлений.}
  \item{campaign_ids}{Числовой вектор, фильтр по рекламным кампаниям.}
  \item{ad_ids}{Числовой вектор, фильтр по рекламным объявлениям.}
  \item{access_token}{Токен доступа к API, полученный с помощью функций vkAuth или vkGetToken}
}
\details{
Если одновременно включены фильтры campaign_ids и ad_ids, то выведены будут следующие объявления:
1. Если id объявления указано в ad_ids, то оно выводится.
2. Если объявление принадлежит рекламной кампании, чей id указан в campaign_ids, то оно выводится.
3. Если существет объявление с id, указанным в ad_ids и оно принадлежит кампании, чей id указан в campaign_ids, то эта рекламная кампания игнорируется, т.е. для нее правило 2 не работает.

Выводится не более 2000 объявлений из не более чем 2000 кампаний. 
}
\value{Дата фрейм со следующими значениями:
\item{1. id}{идентификатор объявления.}
\item{2. campaign_id}{идентификатор кампании.}
\item{3. name}{название объявления.}
\item{4. status}{статус объявления.}
\item{5. ad_format}{формат объявления.}
\item{6. approved}{статус модерации объявления.}
\item{7. all_limit}{общий лимит объявления в рублях. 0 — лимит не задан.}
\item{8. create_time}{дата и время создания объявления.}
\item{9. update_time}{дата и время последнего изменения объявления.}
\item{10. age_restriction}{метка о возрастном ограничении объявления.}
\item{11. category1_id}{ID тематики или подраздела тематики объявления.}
\item{12. category2_id}{ID тематики или подраздела тематики объявления. Дополнительная тематика.}
\item{13. cost_type}{тип оплаты.}
\item{14. cpc}{цена за переход в копейках.}
\item{15. cpm}{цена за 1000 показов в копейках.}
\item{16. impressions_limit}{ ограничение количества показов данного объявления на одного пользователя.}
\item{17. impressions_limited}{признак того, что количество показов объявления на одного пользователя ограничено.} 
\item{18. ad_platform}{рекламные площадки, на которых будет показываться объявление. (если ad_format равен 1, 0 — ВКонтакте и сайты-партнёры, 1 — только ВКонтакте, (если ad_format равен 9) all — все площадки, desktop — полная версия сайта, mobile — мобильный сайт и приложения.)}
\item{19. video}{1 — объявление является видеорекламой.}
}
\author{
Alexey Seleznev
}

\examples{
##Авторизация в вконтакте
my_tok <- vkAuth(app_id = 1,client_secret = "H2Pk8htyFD8024mZaPHm")
#Получаем список объявлений
my_vk_ads <- vkGetAds(account_id = 11111111, 
                      access_token = my_tok$access_token)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.