\name{vkLoadToken}
\alias{vkLoadToken}
\title{Load vkontakte developer token}
\description{vkLoadToken load your developer token, which you need for work with vk.com api }
\usage{
vkLoadToken()
}
\value{R object with vk.com token}
\author{Alexey Seleznev}
\examples{
\dontrun{
##---- Get token from work directory
myToken <- vkLoadToken()
}
}
